requirements.txt
absl-py==1.3.0
adal==1.2.7
adlfs==2022.11.2
aiobotocore==2.4.2
aiohttp==3.8.3
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argcomplete==2.0.0
arrow==1.2.3
async-timeout==4.0.2
asyncssh==2.13.0
atpublic==3.1.1
attrs==22.2.0
azure-core==1.26.1
azure-datalake-store==0.0.52
azure-identity==1.12.0
azure-storage-blob==12.14.1
bcrypt==4.0.1
billiard==3.6.4.0
binaryornot==0.4.4
boto3==1.24.59
botocore==1.27.59
cachetools==5.2.0
celery==5.2.7
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.6
cookiecutter==2.1.1
crcmod==1.7
cryptography==39.0.0
cycler==0.11.0
datasets==2.8.0
decorator==5.1.1
dictdiffer==0.9.0
dill==0.3.6
diskcache==5.4.0
distro==1.8.0
docker-pycreds==0.4.0
dpath==2.1.3
dulwich==0.20.50
dvc==2.38.1
dvc-azure==2.20.5
dvc-data==0.28.4
dvc-gdrive==2.19.1
dvc-gs==2.20.0
dvc-hdfs==2.19.0
dvc-http==2.27.2
dvc-objects==0.14.0
dvc-oss==2.19.0
dvc-render==0.0.15
dvc-s3==2.21.0
dvc-ssh==2.20.0
dvc-studio-client==0.3.0
dvc-task==0.1.8
dvc-webdav==2.19.0
dvc-webhdfs==2.19.0
dvclive==1.3.1
exceptiongroup==1.1.0
filelock==3.9.0
flake8==6.0.0
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
funcy==1.17
future==0.18.2
gcsfs==2022.11.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grandalf==0.6
grpcio==1.51.1
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.2
huggingface-hub==0.11.1
hydra-core==1.3.1
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
iterative-telemetry==0.0.6
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.2.0
kiwisolver==1.4.4
knack==0.10.1
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mccabe==0.7.0
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
multiprocess==0.70.14
nanotime==0.5.2
networkx==2.8.8
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
oss2==2.16.0
ossfs==2023.1.0
packaging==22.0
pandas==1.5.2
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.4.0
pluggy==1.0.0
portalocker==2.6.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pydot==1.4.2
PyDrive2==1.15.0
pyflakes==3.0.1
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.7
pywin32==305
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rich==13.0.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3fs==2022.11.0
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.10.0
scmrepo==0.1.4
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
snakeviz==2.1.1
sniffio==1.3.0
-e git+https://github.com/s185231/MNIST.git@1bedb4b69d00a44959d8a09e58deb4981d82a04f#egg=src
sshfs==2022.6.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.1
torch-tb-profiler==0.4.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
voluptuous==0.13.1
wandb==0.13.7
wcwidth==0.2.5
webdav4==0.9.8
Werkzeug==2.2.2
wget==3.2
wincertstore==0.2
wrapt==1.14.1
xxhash==3.2.0
yarl==1.8.2
zc.lockfile==2.0
